% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/corrvis.R
\name{corrvis}
\alias{corrvis}
\title{corrvis}
\usage{
corrvis(mycorrmatrix)
}
\arguments{
\item{mycorrmatrix}{A correlation matrix created by R's cor function.}
}
\description{
Visualize a correlation matrix using the corrplot library's corrplot function, with color and other options from code in the R Graphics Cookbook by Winston Chang.
}
\examples{
mycors <- cor(df)
corrvis(mycors)
}
\keyword{correlation}

