% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corrviz.R
\name{corrviz}
\alias{corrviz}
\title{corrviz}
\usage{
corrviz(mycorrmatrix)
}
\arguments{
\item{mycorrmatrix}{A correlation matrix created by R's cor function.}
}
\description{
Visualize a correlation matrix using the corrplot library's corrplot function, with color and other options from code in the R Graphics Cookbook by Winston Chang.
}
\examples{
mycors <- cor(df)
corrviz(mycors)
}
\keyword{correlation}

