% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/df_to_xts.R
\name{df_to_xts}
\alias{df_to_xts}
\title{Create an xts object from a common data-frame format.}
\usage{
df_to_xts(mydf)
}
\arguments{
\item{mydf}{A data frame with the first column containg dates, rest of  columns containing data you want. Date column must be in yyyy-mm-dd format, either as character or Date objects.}
}
\description{
Time-series objects can be confusing for the R date/time beginner. This is one way to create an xts object -- class extensible time series (from the zoo package) -- from a data frame containing one column of dates and the rest data.
}
\details{
The resulting xts object can be easily graphed using dygraphs::dygraph(myxtsobject)
}
\examples{
month = c("2014-01-01", "2014-02-01", "2014-03-01", "2014-04-01", "2014-05-01", "2014-06-01", "2014-07-01", "2014-08-01", "2014-09-01", "2014-10-01", "2014-11-01", "2014-12-01")
rate = c(6.6, 6.7, 6.6, 6.2, 6.3, 6.1, 6.2, 6.1, 5.9, 5.7, 5.8, 5.6)
unemp <- data.frame(month, rate, stringsAsFactors = FALSE )

unemp.xts <- df_to_xts(unemp)
# With dygraphs library installed, can then graph this with
# dygraphs::dygraph(unemp.xts)
}
\keyword{series,}
\keyword{time}
\keyword{xts}

