% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grouped_barchart.R
\name{grouped_barchart}
\alias{grouped_barchart}
\title{Create a grouped barchart from data frame}
\usage{
grouped_barchart(mydf, xcolnum, ycolnum, fillcolnum, mytitle = "")
}
\arguments{
\item{mydf}{A data frame
xcolnum integer Data frame column number you want on the x axis
ycolnum integer Data frame column you want on the y axis
fillcolnum integer Data frame column you want bars colored by
mytitle character string optional headline for graph}
}
\description{
Graph format from the R Graphics Cookbook by Winston Chang.
Idea for using ggplot2 within another function from StackOverflow
http://stackoverflow.com/a/27462105/718150
}
\examples{
Initial graph from R Cookbook
http://www.cookbook-r.com/Graphs/Bar_and_line_graphs_(ggplot2)/
dat1 <- data.frame(
sex = factor(c("Female","Female","Male","Male")),
time = factor(c("Lunch","Dinner","Lunch","Dinner"), levels=c("Lunch","Dinner")),
total_bill = c(13.53, 16.81, 16.24, 17.42)
)
p <- grouped_barchart(dat1, 2, 3, 1, "Test graph")
print(p)
}
\keyword{bar}
\keyword{chart}
\keyword{ggplot2,}

