% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/interactive_histogram.R
\name{interactive_histogram}
\alias{interactive_histogram}
\title{interactive_histogram}
\usage{
interactive_histogram(myBasicHistogram, binMin, binMax, slideStep)
}
\arguments{
\item{myBasicHistogram}{a ggiv histogram such as ggvis(mydf, x = ~xcolumn, fill := "gray")
binMin number, the minimum bin width for the histogram
binMax number, the maximum bin width for the histogram
slideStep number, the increment the slider should change with each movement}
}
\description{
Takes a basic ggvis histogram and adds an interactive slider to control bin width.
}
\examples{
initialHistogram <- ggvis(mtcars, x=~mpg, fill := "gray")
 interactive_histogram(initialHistogram, 2, 10, 1)
}
\keyword{ggvis,}
\keyword{histogram}

