% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Draw multiple ggplot2 visualizations on a single plot}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{one or more ggplot2 plot objects}

\item{cols}{integer number of columns across, defaults to 1}
}
\description{
Create your plots and store them in variables, then render in format such as
multiplot(p1, p2, p3, p4, cols=2)
}
\details{
This function is by Winston Chang
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\examples{
\dontrun{
multiplot(p1, p2, p3, p4, cols=2)
}
}
\keyword{ggplot2}

